refactor: rename api files and use date helpers
